requirements.txt
numpy==1.26.4
opencv-contrib-python==4.10.0.84
torch==2.3.1
PyYAML==6.0.1